main update

+++ backend17/src/main/java/com/ocio/backend17/dao/ImagesDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImagesDao extends CrudRepository<Images, Double> {
+public interface ImagesDao extends CrudRepository<Images, Integer> {
 }
